movies/requirements.txt
flask
flask_restful
gunicorn
requests
psycopg2-binary